Split SharedData into ButtonBar and NameSection
