+++ src/client/entity-editor/button-bar/alias-button.js
