+++ src/client/entity-editor/button-bar/identifier-button.js
