+++ src/client/entity-editor/creator.js
@@ -21,17 +21,19 @@ import Immutable from 'immutable';
 import {Provider} from 'react-redux';
 import React from 'react';
 import aliasEditorReducer from './alias-editor/reducer';
+import buttonBarReducer from './button-bar/reducer';
 import {combineReducers} from 'redux-immutable';
 import {createStore} from 'redux';
 import creatorDataReducer from './creator-data/reducer';
 import identifierEditorReducer from './identifier-editor/reducer';
-import sharedDataReducer from './shared-data/reducer';
+import nameSectionReducer from './name-section/reducer';
 
 const rootReducer = combineReducers({
-	sharedData: sharedDataReducer,
+	buttonBar: buttonBarReducer,
 	aliasEditor: aliasEditorReducer,
 	creatorData: creatorDataReducer,
-	identifierEditor: identifierEditorReducer
+	identifierEditor: identifierEditorReducer,
+	nameSection: nameSectionReducer
 });
 
 function generateInitialState(creator, creatorTypes) {
